+++ Tests/SwiftLintFrameworkTests/CommandTests.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable file_length
 import Foundation
 import SourceKittenFramework
 @_spi(TestHelper)
@@ -388,6 +389,24 @@ class CommandTests: XCTestCase {
         )
         XCTAssertEqual(
             violations(Example("print(123)\n// swiftlint:disable:previous nesting\n"), config: configuration),
+            []
+        )
+    }
+
+    func testSuperfluousDisableCommandsDisabledWhenAllRulesDisabled() {
+        XCTAssertEqual(
+            violations(Example("""
+                               // swiftlint:disable all
+                               // swiftlint:disable non_existent_rule_name
+                               """
+            )),
+            []
+        )
+        XCTAssertEqual(
+            violations(Example(
+                "// swiftlint:disable superfluous_disable_command\n" +
+                "// swiftlint:disable non_existent_rule_name\n"
+            )),
             []
         )
     }